+++ src/Render/NeoInputRenderer.php
@@ -104,12 +104,15 @@ class NeoInputRenderer
     public function selectBox($el, array $attrs, array $options): string
     {
         $prompt = $el instanceof SelectBox ? $el->getPrompt() : null;
-
+        /** @var \Efabrica\NeoForms\Control\SelectBox $el */
         return $this->block('selectBox', [
             'attrs' => $attrs,
             'options' => $el->getOptions() + $options,
             'prompt' => $prompt,
-            'items' => $el->getItems(),
+
+            /** @see \Efabrica\NeoForms\Control\SelectBoxTrait */
+            'items' => $el->getItems(true),
+
             'selected' => $el->getValue(),
         ]);
     }